feat: add hardhat tasks to interact with the example contract
