+++ via-playground/scripts/tasks/index.ts
@@ -1,0 +1,4 @@
+export * from './contribute';
+export * from './withdrawFunds';
+export * from './deploy';
+export * from './view';
